Added new stock system
The concept of StockMovement was implemented, an object that stores information about orders, products and related quantities along with whether it is an IN or OUT movement.
OUT movements will be generated when creating an order or modifying it by increasing quantities.
IN movements will be generated when eliminating an order or modifying it by decreasing the quantities.
The stock of the products will be updated based on these movements.